Move models namespaces
