+++ app/Monitor/Models/Config/UpdateConfigModel.cs
@@ -1,4 +1,4 @@
-namespace Monitor.Models;
+namespace Monitor.Models.Config;
 
 public class UpdateConfigModel
 {
