+++ app/Monitor/Models/Emails/SummaryReport.cs
@@ -1,3 +1,6 @@
+using Monitor.Shared.Models.Instances;
+using Monitor.Shared.Models.Reports;
+
 namespace Monitor.Models.Emails;
 
 public record SummaryReport(List<ReportModel> ConflictingSites, List<ReportModel> ConflictingNames,
