+++ app/Monitor/Services/StudyService.cs
@@ -9,7 +9,7 @@ using Monitor.Config;
 using Monitor.Shared.Constants;
 using Monitor.Data.Constants;
 using Monitor.Data.Entities;
-using Monitor.Models.Studies;
+using Monitor.Shared.Models.Studies;
 using StudyUser = Monitor.Data.Entities.StudyUser;
 
 namespace Monitor.Services;
