+++ lib/Monitor.Shared/Models/Instances/InstanceModel.cs
@@ -1,6 +1,6 @@
-namespace Monitor.Models;
+namespace Monitor.Shared.Models.Instances;
 
-public class ReportStatusModel
+public class InstanceModel
 {
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
