+++ lib/Monitor.Shared/Models/Reports/ReportTypeModel.cs
@@ -1,6 +1,6 @@
-namespace Monitor.Models;
+namespace Monitor.Shared.Models.Reports;
 
-public class InstanceModel
+public class ReportTypeModel
 {
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
