+++ lib/Monitor.Shared/Models/Sites/SiteModel.cs
@@ -1,4 +1,4 @@
-namespace Monitor.Models;
+namespace Monitor.Shared.Models.Sites;
 
 public class SiteModel
 {
